this commit will show kata 11 completed and edits to my controller and model files
