+++ Controllers/comments-controllers.js
@@ -1,4 +1,19 @@
-const {removeComment} = require ('../Models/comment-models.js')
+const {removeComment, insertNewComment, editVotes} = require ('../Models/comment-models.js')
+
+
+exports.postNewComment = (req, res, next) => {
+    
+    const {article_id} = req.params
+    const newComment = req.body.body
+    const author = req.body.username
+insertNewComment(newComment, article_id, author)
+.then((newComment)=>{
+res.status(201).send({yourNewComment: newComment[0].body})
+})
+.catch((err) => {
+    next(err)
+})
+}
 
 exports.deleteComment = (req, res, next) => {
     const commentId = req.params.comment_id
@@ -8,3 +23,15 @@ exports.deleteComment = (req, res, next) => {
         next(err)
     })
     }
+
+ exports.patchVotes = (req, res, next) => {
+        const {article_id} = req.params
+      const incVoteNum = req.body.inc_votes
+      editVotes(article_id, incVoteNum).then((editedArticle)=>{
+          res.status(200).send({updatedArticle: editedArticle[0] })
+      })
+      .catch((err) => {
+          next(err)
+      })
+      }
+
