+++ Models/users-models.js
@@ -1,0 +1,11 @@
+const db = require('../db/connection.js')
+
+exports.fetchAllUsers = () =>{
+    return db.query(`SELECT * FROM users;`).then(({rows})=>{
+      
+     return rows
+    })
+  }
+  
+      
+
