+++ __tests__/app.test.js
@@ -272,6 +272,32 @@ describe('POST /api/articles/:article_id/comments', () => {
 
 
 
+describe('DELETE/api/comments/:comments_id', () => {
+    it('returns a 204 status code and no content', () => {
+        return request(app)
+            .delete('/api/comments/4')
+            .expect(204)
+            .then(({ body }) => {
+                expect(body).toEqual({})
+            })
+    })
+    it('DELETE: returns 404 status code if tries to delete a comment_id that does not exist', () => {
+        return request(app)
+            .delete('/api/comments/11111')
+            .expect(404)
+            .then(({ body }) => {
+                expect(body.msg).toBe('comment_id does not exist')
+            })
+    })
+    it('DELETE: returns 404 status code if tries to delete a comment_id that does not exist', () => {
+        return request(app)
+            .delete('/api/comments/not-a-comment-id')
+            .expect(400)
+            .then(({ body }) => {
+                expect(body.msg).toBe('bad request')
+            })
+    })
+})
 describe('PATCH /api/articles/article_id', () => {
     it('will return a 200 status code and an updated article with the vote count incremented', () => {
         const votePatch = { inc_votes: 20 }
@@ -292,54 +318,6 @@ describe('PATCH /api/articles/article_id', () => {
                 }
                 expect(body).toMatchObject({ updatedArticle: desiredObj })
             })
-
-
-describe('DELETE/api/comments/:comments_id', () => {
-    it('returns a 204 status code and no content', () => {
-        return request(app)
-       .delete('/api/comments/4')
-       .expect(204)
-       .then(({body}) =>{
-        expect(body).toEqual({})
-       })
-    })
-    it('DELETE: returns 404 status code if tries to delete a comment_id that does not exist', ()=>{ 
-    return request(app)
-    .delete('/api/comments/11111')
-    .expect(404)
-    .then(({body})=>{
-        expect(body.msg).toBe('comment_id does not exist')
-    })
-    })
-    it('DELETE: returns 404 status code if tries to delete a comment_id with an invalid comment_id', ()=>{ 
-        return request(app)
-        .delete('/api/comments/not-a-comment-id')
-        .expect(400)
-        .then(({body})=>{
-            expect(body.msg).toBe('bad request')
-        })
-})
-  
-describe('PATCH /api/articles/article_id', () =>{
-    it('will return a 200 status code and an updated article with the vote count incremented', ()=>{
-    const votePatch = { inc_votes: 20 }
-    return request(app)
-    .patch('/api/articles/4')
-    .send(votePatch)
-    .expect(200)
-    .then(({body})=>{
-        const desiredObj = {
-            title: "Student SUES Mitch!",
-            topic: "mitch",
-            author: "rogersop",
-            body: "We all love Mitch and his wonderful, unique typing style. However, the volume of his typing has ALLEGEDLY burst another students eardrums, and they are now suing for damages",
-            created_at: "2020-05-06T01:14:00.000Z",
-            votes: 20,
-            article_img_url:
-              "https://images.pexels.com/photos/158651/news-newsletter-newspaper-information-158651.jpeg?w=700&h=700",
-          }
-        expect(body).toMatchObject({updatedArticle: desiredObj})
-
     })
 
 
@@ -388,8 +366,6 @@ describe('PATCH /api/articles/article_id', () =>{
 })
 
 
-
-
 describe('GET /api/users', () => {
     it('should return an array of user objects', () => {
         return request(app)
@@ -408,5 +384,36 @@ describe('GET /api/users', () => {
     })
 })
 
-})
+describe('GET /api/articles?topic=topicname', ()=>{
+    it('GET: should filter the articles and only return the articles that include the topic specified', ()=>{
+        return request(app)
+        .get('/api/articles?topic=mitch')
+        .expect(200)
+        .then(({body})=> {
+            expect(body.articles.length).toBe(12)
+            body.articles.forEach((article)=>{
+                expect(article.topic).toBe('mitch')
+            })
+        })
+    })
+    it('GET: should return 200 and an empty array if given valid topic but there are no articles associated with this topic', ()=>{
+        return request(app)
+        .get('/api/articles?topic=paper')
+        .expect(200)
+        .then(({body})=> {
+            expect(body.articles.length).toBe(0)
+                expect(body.articles).toEqual([])
+            })
+        })
+    
+    it('GET: should return 404 not found if user inputs topic which does not exist', ()=>{
+        return request(app)
+        .get('/api/articles?topic=not-a-topic')
+        .expect(404)
+        .then(({body})=> {
+                expect(body.msg).toBe('invalid topic')
+            })
+        })
+    })
+
