+++ app.js
@@ -1,13 +1,11 @@
 const express = require("express");
 const data = require('./db/data/test-data/index.js')
 
-const {postNewComment, patchVotes} = require ('./Controllers/post-and-patch-controllers.js')
+const {getAllUsers} = require ('./Controllers/users-controllers.js')
+const { getAllEndpoints, getArticleById, getAllArticles, getArticleCommentsById} = require('./Controllers/articles-controllers.js')
+const { getAllTopics} = require('./Controllers/topics-controllers.js')
+const {deleteComment, postNewComment, patchVotes} = require ('./Controllers/comments-controllers.js')
 
-const { getAllTopics, getAllEndpoints, getArticleById, getAllArticles, getArticleCommentsById, getAllUsers} = require('./Controllers/get-controllers.js')
-
-
-const { getAllTopics, getAllEndpoints, getArticleById, getAllArticles, getArticleCommentsById } = require('./Controllers/get-controllers.js')
-const {deleteComment} = require ('./Controllers/comments-controllers.js')
 
 const { handleCustomErrors, handleSQLErrors } = require('./Error handling/error-handling.js');
 
@@ -21,11 +19,11 @@ app.get('/api/articles/:article_id', getArticleById)
 app.get('/api/articles', getAllArticles)
 app.get('/api/articles/:article_id/comments', getArticleCommentsById)
 app.get('/api/users', getAllUsers)
+
+
 app.post('/api/articles/:article_id/comments', postNewComment)
 
-
 app.delete('/api/comments/:comment_id', deleteComment)
-
 
 app.patch('/api/articles/:article_id', patchVotes)
